cast hidden to int

+++ hyper_search.py
@@ -229,7 +229,7 @@ def objective(trial):
         if i == 0:
             model.add(LSTM(hidden_units, return_sequences=num_layers > 1, input_shape=(num_previous_intervals, num_features)))
         else:
-            model.add(LSTM(hidden_units*layer_multiplier*i, return_sequences=i < num_layers - 1))
+            model.add(LSTM(int(hidden_units*layer_multiplier*i), return_sequences=i < num_layers - 1))
         model.add(Dropout(dropout_rate))
     model.add(Dense(100))
